requirements.txt
asgiref==3.5.2
Django==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2021.1
sqlparse==0.4.1
